requirements.txt
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
